only advertise eden support at the end of extsetup()
Summary:
Update the Eden mercurial extension to only update the repository class to
indicate that it supports the `eden` requirement at the very end of
`extsetup()`.

Previously we added this capability as soon as the extension module was first
imported.  If an exception occurred after this point but before `extsetup()`
completes mercurial would try to proceed anyway without the Eden extension.
However, we really should bail out at this point if the repository requires
Eden rather than falling back and letting the normal dirstate code try to deal
with the Eden dirstate file.

Reviewed By: chadaustin

Differential Revision: D6746819

fbshipit-source-id: 0faa86e5fac58636014b44aa8277655ae94208a4