suppress a pyre error about mercurial vs edenscm.mercurial
Summary:
pyre complains since it cannot find the `mercurial` package.
We intentionally want to keep using the old name until the package change in
Mercurial is fully rolled out, so suppress this warning for now.

This module itself will likely just be deleted in the future, as it has been
integrated into mercurial itself for recent mercurial releases.

Reviewed By: quark-zju

Differential Revision: D13858170

fbshipit-source-id: a12c73e54da9afc63e136af9922dd6a2cf6912be